Move load() into a separate file
