+++ app/tests/src/config/test_config.py
@@ -6,10 +6,11 @@ import pydantic.error_wrappers
 import pytest
 
 from src import config
+from src.config import load
 
 
 def test_load_with_override():
-    conf = config.load(
+    conf = load.load(
         environment_name="local", environ={"app_host": "test.host", "app_port": 999, "port": 888}
     )
 
@@ -20,11 +21,11 @@ def test_load_with_override():
 
 def test_load_with_invalid_override():
     with pytest.raises(pydantic.error_wrappers.ValidationError):
-        config.load(environment_name="local", environ={"app_port": "not_a_number"})
+        load.load(environment_name="local", environ={"app_port": "not_a_number"})
 
 
 def test_load():
-    conf = config.load(environment_name="local")
+    conf = load.load(environment_name="local")
 
     assert isinstance(conf, config.RootConfig)
     assert conf.app.host == "127.0.0.1"
@@ -32,14 +33,14 @@ def test_load():
 
 def test_load_invalid_environment_name():
     with pytest.raises(ModuleNotFoundError):
-        config.load(environment_name="does_not_exist")
+        load.load(environment_name="does_not_exist")
 
 
 def test_load_all():
     """This test is important to confirm that all configurations are valid - otherwise we would
     not know until runtime in the appropriate environment."""
 
-    all_configs = config.load_all()
+    all_configs = load.load_all()
 
     # We expect at least these configs to exist - there may be others too.
     assert all_configs.keys() >= {"local", "dev", "prod"}
